Move exe modules to main library
